Moved files to match new and hopefully working file structure.
Partially cherry-picked from 42e4ffcf915c33d6d803406d8f520c4ceeed2b04.
