+++ lib/kramdown_html.rb
@@ -1,5 +1,5 @@
 require 'kramdown'
-require "swedbank/pay/jekyll/plantuml/plantuml_converter"
+require "swedbank-pay-jekyll-plantuml/plantuml_converter"
 
 class Kramdown::Converter::Html
     alias_method :super_convert_codeblock, :convert_codeblock
@@ -9,7 +9,7 @@ class Kramdown::Converter::Html
             return super_convert_codeblock(element, indent)
         end
 
-        converter = Swedbank::Pay::Jekyll::Plantuml::Converter.new
+        converter = SwedbankPayJekyllPlantuml::Converter.new
         return converter.convert_plantuml_to_svg(element.value)
     end
 end
