+++ lib/swedbank-pay-jekyll-plantuml.rb
