--- lib/swedbank/pay/jekyll/plantuml/version.rb
@@ -1,9 +1,0 @@
-module Swedbank
-  module Pay
-    module Jekyll
-      module Plantuml
-        VERSION = "0.2.0"
-      end
-    end
-  end
-end
